Adding Test Cases for RewardService and Modifying the Double Value for Rewards Calculation

+++ src/main/java/com/customer_rewards/rewards_calculation/dto/TransactionResponseDto.java
@@ -14,7 +14,7 @@ public class TransactionResponseDto {
 
     private Date PurchaseDate;
 
-    private Integer rewardPoints;
+    private Double rewardPoints;
 
     @JsonProperty("customerDetail")
     private CustomerResponseDto customerResponseDto;
